copilot-local-agent/eval/requirements.txt
azure-ai-evaluation>=1.0.0
azure-ai-projects>=2.0.0
azure-monitor-query>=1.0.0
azure-identity>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
